requirements-prod.txt
# Production-specific dependencies for fraud detection system

# Web server
gunicorn==21.2.0
uvicorn[standard]==0.24.0

# Production monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-instrumentation-redis==0.42b0
opentelemetry-instrumentation-kafka-python==0.42b0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-prometheus==1.12.0rc1

# Security
cryptography==41.0.7
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Performance optimization
uvloop==0.19.0
httptools==0.6.1

# Health checks and monitoring
psutil==5.9.6

# Production database drivers
psycopg2-binary==2.9.9

# Caching
hiredis==2.2.3

# Configuration management
python-dotenv==1.0.0

# Process management
supervisor==4.2.5

# Memory profiling (optional)
memory-profiler==0.61.0

# Production-grade JSON logging
python-json-logger==2.0.7

# Rate limiting
slowapi==0.1.9

# CORS handling
fastapi-cors==0.0.6

# Request ID tracking
fastapi-request-id==1.0.0

# Graceful shutdown
fastapi-graceful-shutdown==0.2.0